made forms into components
